Refactored personnel to use service layer

+++ backend/api/src/routes/modules/common.rs
@@ -43,8 +43,8 @@ impl From<User> for RoleResponse {
 
 #[derive(Debug, Deserialize)]
 pub struct RoleQuery {
-    pub page: Option<u32>,
-    pub per_page: Option<u32>,
+    pub page: Option<u64>,
+    pub per_page: Option<u64>,
     pub query: Option<String>,
     pub email: Option<String>,
     pub username: Option<String>,
@@ -54,8 +54,8 @@ pub struct RoleQuery {
 #[derive(serde::Serialize)]
 pub struct PaginatedRoleResponse {
     pub users: Vec<RoleResponse>,
-    pub page: u32,
-    pub per_page: u32,
+    pub page: u64,
+    pub per_page: u64,
     pub total: u64,
 }
